Fix Subscription Lock
