+++ crates/bcr-wallet-cli/bob.toml
@@ -5,7 +5,7 @@ esplora_base_urls = ["https://esplora.minibill.tech", "https://blockstream.info"
 [wallets.49b81beaed5cdd051ff5070906be2d8c17508ddb01436f5d6e56216d6f634c0c]
 network = "testnet"
 nostr_relays = ["wss://relay.wildcat0.clowder-dev.minibill.tech", "wss://relay.wildcat1.clowder-dev.minibill.tech", "wss://relay.wildcat3.clowder-dev.minibill.tech"]
-mint_url = "https://mint.wildcat3.clowder-dev.minibill.tech"
+mint_url = "https://mint.wildcat1.clowder-dev.minibill.tech"
 mnemonic = "struggle toilet attitude trim obey grab alter apology across bamboo palm naive"
 
 [wallets.9b0ade58c90766fc6c4313dde1769b1146a60e0f6407919fbd41053e93c78030]
